+++ rust/src/frb_generated.rs
@@ -434,6 +434,7 @@ fn wire__crate__api__typst__TypstEngine_compile_impl(
             let api_sys_time = <Option<i64>>::sse_decode(&mut deserializer);
             let api_inputs =
                 <Option<std::collections::HashMap<String, String>>>::sse_decode(&mut deserializer);
+            let api_allow_packages = <bool>::sse_decode(&mut deserializer);
             deserializer.end();
             move |context| {
                 transform_result_sse::<_, crate::api::typst::TypstCompileError>((move || {
@@ -457,6 +458,7 @@ fn wire__crate__api__typst__TypstEngine_compile_impl(
                         api_files,
                         api_sys_time,
                         api_inputs,
+                        api_allow_packages,
                     )?;
                     Ok(output_ok)
                 })(
@@ -653,6 +655,13 @@ impl SseDecode for String {
     fn sse_decode(deserializer: &mut flutter_rust_bridge::for_generated::SseDeserializer) -> Self {
         let mut inner = <Vec<u8>>::sse_decode(deserializer);
         return String::from_utf8(inner).unwrap();
+    }
+}
+
+impl SseDecode for bool {
+    // Codec=Sse (Serialization based), see doc to use other codecs
+    fn sse_decode(deserializer: &mut flutter_rust_bridge::for_generated::SseDeserializer) -> Self {
+        deserializer.cursor.read_u8().unwrap() != 0
     }
 }
 
@@ -920,13 +929,6 @@ impl SseDecode for crate::api::typst::VirtualFile {
             path: var_path,
             bytes: var_bytes,
         };
-    }
-}
-
-impl SseDecode for bool {
-    // Codec=Sse (Serialization based), see doc to use other codecs
-    fn sse_decode(deserializer: &mut flutter_rust_bridge::for_generated::SseDeserializer) -> Self {
-        deserializer.cursor.read_u8().unwrap() != 0
     }
 }
 
@@ -1208,6 +1210,13 @@ impl SseEncode for String {
     }
 }
 
+impl SseEncode for bool {
+    // Codec=Sse (Serialization based), see doc to use other codecs
+    fn sse_encode(self, serializer: &mut flutter_rust_bridge::for_generated::SseSerializer) {
+        serializer.cursor.write_u8(self as _).unwrap();
+    }
+}
+
 impl SseEncode for f32 {
     // Codec=Sse (Serialization based), see doc to use other codecs
     fn sse_encode(self, serializer: &mut flutter_rust_bridge::for_generated::SseSerializer) {
@@ -1427,13 +1436,6 @@ impl SseEncode for crate::api::typst::VirtualFile {
     fn sse_encode(self, serializer: &mut flutter_rust_bridge::for_generated::SseSerializer) {
         <String>::sse_encode(self.path, serializer);
         <Vec<u8>>::sse_encode(self.bytes, serializer);
-    }
-}
-
-impl SseEncode for bool {
-    // Codec=Sse (Serialization based), see doc to use other codecs
-    fn sse_encode(self, serializer: &mut flutter_rust_bridge::for_generated::SseSerializer) {
-        serializer.cursor.write_u8(self as _).unwrap();
     }
 }